refactor to Original, simplify layout enum
