+++ src/android/app/src/main/res/menu/menu_landscape_screen_layout.xml
@@ -20,8 +20,8 @@
             android:title="@string/emulation_screen_layout_hybrid" />
 
         <item
-            android:id="@+id/menu_screen_layout_above_below"
-            android:title="@string/emulation_screen_layout_above_below" />
+            android:id="@+id/menu_screen_layout_original"
+            android:title="@string/emulation_screen_layout_original" />
 
         <item
             android:id="@+id/menu_screen_layout_custom"
